Refactor the genauth code.
This splits the AST parsing code from the C++ generator code, and it
splits both from the genauth application. It also performs extensive
function refactoring in the C++ generator code.
